Refactor contribution form components and remove HelpTooltip
- Deleted the HelpTooltip component as it was deemed unnecessary.
- Replaced instances of HelpTooltip in the OsblDataSheet and OsblHeader components with a shared version from the new location.
- Refactored the InterventionAreaAsyncCreatableSelect and KeywordAsyncCreatableSelect components to utilize MyAsyncCreatableSelect for improved code reuse and maintainability.
- Updated database schema constraints to allow non-negative employee counts, enhancing data integrity.

This commit streamlines the contribution form by consolidating tooltip functionality and improving component structure.
