+++ app/frontend/components/shared/MyAsyncCreatableSelect.tsx
@@ -1,0 +1,102 @@
+import { ReactElement, useState, useCallback } from 'react'
+import AsyncCreatableSelect from 'react-select/async-creatable'
+import axios from 'axios'
+import debounce from 'lodash.debounce'
+import { FormProps } from '@/pages/Contribution/types'
+
+interface Option {
+  readonly label: string
+  readonly value: string
+}
+
+interface MyAsyncCreatableSelectProps extends Pick<FormProps, 'data' | 'setData'> {
+  attributeName: string
+  endpoint: string
+  minInputLength?: number
+  placeholder: string
+}
+
+const createOption = (data: { id: number, name: string }): Option => ({
+  label: data.name,
+  value: String(data.id)
+})
+
+export default function MyAsyncCreatableSelect({
+  data,
+  setData,
+  attributeName,
+  endpoint,
+  minInputLength = 2,
+  placeholder
+}: MyAsyncCreatableSelectProps): ReactElement {
+  const [isLoading, setIsLoading] = useState(false)
+  const [value, setValue] = useState<Option[]>([])
+
+  const fetchOptions = async (inputValue: string): Promise<Option[]> => {
+    if (inputValue.length < minInputLength) return []
+
+    const response = await fetch(`/${endpoint}?query=${encodeURIComponent(inputValue)}`)
+    const data = await response.json()
+    return data.map((data: { id: number, name: string }) => createOption(data))
+  }
+
+  const debouncedFetchOptions = useCallback(
+    debounce((inputValue: string, callback: (options: Option[]) => void) => {
+      void fetchOptions(inputValue).then(callback)
+    }, 300),
+    []
+  )
+
+  const promiseOptions = async (inputValue: string): Promise<Option[]> => {
+    return await new Promise((resolve) => {
+      debouncedFetchOptions(inputValue, resolve)
+    })
+  }
+
+  const handleCreate = (inputValue: string): void => {
+    void (async () => {
+      setIsLoading(true)
+      try {
+        const csrfToken = document.querySelector('meta[name="csrf-token"]')?.getAttribute('content')
+        const response = await axios.post(
+          `/${endpoint}`,
+          { name: inputValue },
+          {
+            headers: {
+              'X-CSRF-Token': csrfToken
+            }
+          }
+        )
+        const newOption = createOption(response.data)
+        setValue((prev) => [...prev, newOption])
+        setData(attributeName, [...data[attributeName], { [`${endpoint.slice(0, -1)}_id`]: newOption.value }])
+      } finally {
+        setIsLoading(false)
+      }
+    })()
+  }
+
+  return (
+    <AsyncCreatableSelect
+      isMulti
+      cacheOptions
+      value={value}
+      loadOptions={promiseOptions}
+      onCreateOption={handleCreate}
+      onChange={(value) => {
+        console.log('value', value)
+        setValue(value as Option[])
+        setData(attributeName, value.map(option => ({
+          [`${endpoint.slice(0, -1)}_id`]: option.value
+        })))
+      }}
+      isDisabled={isLoading}
+      isLoading={isLoading}
+      classNames={{
+        multiValueLabel: () => 'bg-secondary',
+        multiValueRemove: () => 'bg-secondary'
+      }}
+      placeholder={placeholder}
+    />
+  )
+}
